removed duplicate enums from descriptor package

+++ olca-ilcd/src/main/java/org/openlca/ilcd/descriptors/ClassType.java
@@ -2,7 +2,6 @@
 package org.openlca.ilcd.descriptors;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import javax.xml.bind.annotation.XmlAttribute;
 import javax.xml.bind.annotation.XmlValue;
@@ -15,6 +14,6 @@ public class ClassType implements Serializable {
 	public String content;
 
 	@XmlAttribute(name = "level", required = true)
-	public BigInteger level;
+	public int level;
 
 }
