+++ olca-ilcd/src/main/java/org/openlca/ilcd/descriptors/Time.java
@@ -2,7 +2,6 @@
 package org.openlca.ilcd.descriptors;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -19,7 +18,7 @@ public class Time implements Serializable {
 
 	private final static long serialVersionUID = 1L;
 
-	public BigInteger referenceYear;
-	public BigInteger validUntil;
+	public Integer referenceYear;
+	public Integer validUntil;
 
 }
